Upgrade boot 2.0.0 (#93)
* upgrade starter

* Upgrading to boot 2.0.0. Replacing spring-boot-starter-data-elasticsearch by a bare 2.4.6 elasticsearch

* Change message properties encoding to utf-8

* Test fixing

* Merge from master

* Fixing tests classes

* Removing html from properties


* Elasticsearch configuration

* Fixes #80

* Fixes #65


* Fixes #22

* Fixes #87

* Fixes #87

* Fixes #47


* Fixes #41

* Fixes #42

* Moving all templates and public resources to a more isolated location

* Remove old resources


* Project generator optimizations

* Add kotlin source language for project initializr (#92)

* Add source language to initializr options

* Redirect to /initializr on /

* Add Application.kt generation when Kotlin language is selected using KotlinPoet

* Add kotlin configuration to the pom.xml when using Kotlin or Java+Kotlin

* Add kotlin project initializr with WAR deployment

* Camunda autoconfiguration

* Fixes #90

* Putting documentation in the Initializr app

* Upgraded to spring boot 2.0.0.RC1

+++ blossom-core/blossom-core-common/src/main/java/fr/blossom/core/common/search/SearchEngineImpl.java
@@ -100,7 +100,7 @@ public class SearchEngineImpl<DTO extends AbstractDTO> implements SearchEngine {
     }
 
     SearchRequestBuilder searchRequest = this.client.prepareSearch(this.configuration.getAlias()).setQuery(query)
-      .setSize(pageable.getPageSize()).setFrom(pageable.getOffset());
+      .setSize(pageable.getPageSize()).setFrom((int) pageable.getOffset());
 
     Sort sort = pageable.getSort();
     if (sort != null) {
